This should improve performance by removing the need to have a reference to the patient session when performing the join, which prevents Rails from being able to run a single query to avoid N+1 issues. This does slightly break functionality, but since we're only running with one programme at the moment, and other functionality is broken with regards to multiple programmes, it seems like a reasonable fix for now.