@thomasleese 2024-11-05
Locally this is a big performance improvement (from @benilovj):

```
before: 08:19:43 web.1  | Completed 200 OK in 39626ms (Views: 140.9ms | ActiveRecord: 2530.5ms (792 queries, 1 cached) | GC: 6163.8ms)

after: 08:20:56 web.1  | Completed 200 OK in 10149ms (Views: 95.0ms | ActiveRecord: 1837.5ms (23 queries, 1 cached) | GC: 1785.4ms)
```

In the future we'll likely need to look at caching these statistics values or patient session stats as we will need to filter on programmes, but it's also possible that the data model will change in a way that avoids the expensive queries in the future anyway.